added unflatten dict tests
